firesim target decoupling
