+++ sim/src/main/scala/firesim/ExampleTarget.scala
@@ -30,7 +30,7 @@ import java.io.File
 * determine which driver to build.
 *******************************************************************************/
 
-class FireSim(implicit p: Parameters) extends RocketSubsystem
+class FireSimExample(implicit p: Parameters) extends RocketSubsystem
     with CanHaveMisalignedMasterAXI4MemPort
     with HasPeripheryBootROM
 //    with HasSystemErrorSlave
@@ -43,9 +43,7 @@ class FireSim(implicit p: Parameters) extends RocketSubsystem
 {
   val hasTraces = rocketTiles.map(_.rocketParams.trace).reduce(_ || _)
 
-  override lazy val module =
-    if (hasTraces) new FireSimModuleImpTraced(this)
-    else new FireSimModuleImp(this)
+  override lazy val module = new FireSimExampleModuleImp(this)
 
   // Error device used for testing and to NACK invalid front port transactions
   val error = LazyModule(new TLError(p(ErrorDeviceKey), sbus.beatBytes))
@@ -53,7 +51,7 @@ class FireSim(implicit p: Parameters) extends RocketSubsystem
   sbus.coupleTo("slave_named_error"){ error.node := TLBuffer() := _ } 
 }
 
-class FireSimModuleImp[+L <: FireSim](l: L) extends RocketSubsystemModuleImp(l)
+class FireSimExampleModuleImp[+L <: FireSim](l: L) extends RocketSubsystemModuleImp(l)
     with HasRTCModuleImp
     with CanHaveMisalignedMasterAXI4MemPortModuleImp
     with HasPeripheryBootROMModuleImp
@@ -64,10 +62,7 @@ class FireSimModuleImp[+L <: FireSim](l: L) extends RocketSubsystemModuleImp(l)
     with HasPeripheryIceNICModuleImpValidOnly
     with HasPeripheryBlockDeviceModuleImp
 
-class FireSimModuleImpTraced[+L <: FireSim](l: L) extends FireSimModuleImp(l)
-    with CanHaveRocketTraceIO
-
-class FireSimNoNIC(implicit p: Parameters) extends RocketSubsystem
+class FireSimNoNICExample(implicit p: Parameters) extends RocketSubsystem
     with CanHaveMisalignedMasterAXI4MemPort
     with HasPeripheryBootROM
 //    with HasSystemErrorSlave
@@ -77,11 +72,8 @@ class FireSimNoNIC(implicit p: Parameters) extends RocketSubsystem
     with HasPeripheryUART
     with HasPeripheryBlockDevice
 {
-  val hasTraces = rocketTiles.map(_.rocketParams.trace).reduce(_ || _)
 
-  override lazy val module =
-    if (hasTraces) new FireSimNoNICModuleImpTraced(this)
-    else new FireSimNoNICModuleImp(this)
+  override lazy val module = new FireSimNoNICExampleModuleImp(this)
 
   // Error device used for testing and to NACK invalid front port transactions
   val error = LazyModule(new TLError(p(ErrorDeviceKey), sbus.beatBytes))
@@ -89,7 +81,7 @@ class FireSimNoNIC(implicit p: Parameters) extends RocketSubsystem
   sbus.coupleTo("slave_named_error"){ error.node := TLBuffer() := _ } 
 }
 
-class FireSimNoNICModuleImp[+L <: FireSimNoNIC](l: L) extends RocketSubsystemModuleImp(l)
+class FireSimNoNICExampleModuleImp[+L <: FireSimNoNIC](l: L) extends RocketSubsystemModuleImp(l)
     with HasRTCModuleImp
     with CanHaveMisalignedMasterAXI4MemPortModuleImp
     with HasPeripheryBootROMModuleImp
@@ -99,80 +91,6 @@ class FireSimNoNICModuleImp[+L <: FireSimNoNIC](l: L) extends RocketSubsystemMod
     with HasPeripheryUARTModuleImp
     with HasPeripheryBlockDeviceModuleImp
 
-class FireSimNoNICModuleImpTraced[+L <: FireSimNoNIC](l: L) extends FireSimNoNICModuleImp(l)
-    with CanHaveRocketTraceIO
-
-class FireBoom(implicit p: Parameters) extends BoomSubsystem
-    with CanHaveMisalignedMasterAXI4MemPort
-    with HasPeripheryBootROM
-//    with HasSystemErrorSlave
-    // with HasSyncExtInterrupts
-    with HasNoDebug
-    with HasPeripherySerial
-    with HasPeripheryUART
-    with HasPeripheryIceNIC
-    with HasPeripheryBlockDevice
-{
-  val hasTraces = boomTiles.map(_.boomParams.trace).reduce(_ || _)
-
-  override lazy val module = 
-    if (hasTraces) new FireBoomModuleImpTraced(this)
-    else new FireBoomModuleImp(this)
-
-  // Error device used for testing and to NACK invalid front port transactions
-  val error = LazyModule(new TLError(p(ErrorDeviceKey), sbus.beatBytes))
-  // always buffer the error device because no one cares about its latency
-  sbus.coupleTo("slave_named_error"){ error.node := TLBuffer() := _ } 
-}
-
-class FireBoomModuleImp[+L <: FireBoom](l: L) extends BoomSubsystemModule(l)
-    with HasRTCModuleImp
-    with CanHaveMisalignedMasterAXI4MemPortModuleImp
-    with HasPeripheryBootROMModuleImp
-    // with HasExtInterruptsModuleImp
-    with HasNoDebugModuleImp
-    with HasPeripherySerialModuleImp
-    with HasPeripheryUARTModuleImp
-    with HasPeripheryIceNICModuleImpValidOnly
-    with HasPeripheryBlockDeviceModuleImp
-
-class FireBoomModuleImpTraced[+L <: FireBoom](l: L) extends FireBoomModuleImp(l)
-    with CanHaveBoomTraceIO
-
-class FireBoomNoNIC(implicit p: Parameters) extends BoomSubsystem
-    with CanHaveMisalignedMasterAXI4MemPort
-    with HasPeripheryBootROM
-//    with HasSystemErrorSlave
-    // with HasSyncExtInterrupts
-    with HasNoDebug
-    with HasPeripherySerial
-    with HasPeripheryUART
-    with HasPeripheryBlockDevice
-{
-  val hasTraces = boomTiles.map(_.boomParams.trace).reduce(_ || _)
-
-  override lazy val module = 
-    if (hasTraces) new FireBoomNoNICModuleImpTraced(this)
-    else new FireBoomNoNICModuleImp(this)
-
-  // Error device used for testing and to NACK invalid front port transactions
-  val error = LazyModule(new TLError(p(ErrorDeviceKey), sbus.beatBytes))
-  // always buffer the error device because no one cares about its latency
-  sbus.coupleTo("slave_named_error"){ error.node := TLBuffer() := _ } 
-}
-
-class FireBoomNoNICModuleImp[+L <: FireBoomNoNIC](l: L) extends BoomSubsystemModule(l)
-    with HasRTCModuleImp
-    with CanHaveMisalignedMasterAXI4MemPortModuleImp
-    with HasPeripheryBootROMModuleImp
-    // with HasExtInterruptsModuleImp
-    with HasNoDebugModuleImp
-    with HasPeripherySerialModuleImp
-    with HasPeripheryUARTModuleImp
-    with HasPeripheryBlockDeviceModuleImp
-
-class FireBoomNoNICModuleImpTraced[+L <: FireBoomNoNIC](l: L) extends FireBoomNoNICModuleImp(l)
-    with CanHaveBoomTraceIO
 
 case object NumNodes extends Field[Int]
 
@@ -217,4 +135,3 @@ class FireSimSupernode(implicit p: Parameters) extends Module {
     n.debug.clockeddmi.get.dmi.req.bits.op := DontCare
   } }
 }
-
