+++ sim/src/main/scala/firesim/ExampleTargetConfigs.scala
@@ -1,0 +1,55 @@
+package firesim
+
+import freechips.rocketchip.config.{Parameters, Config}
+import freechips.rocketchip.tile._
+import freechips.rocketchip.tilelink._
+import freechips.rocketchip.subsystem._
+import firechip._
+
+//**********************************************************************************
+//* Supernode Configurations
+//*********************************************************************************/
+class WithNumNodes(n: Int) extends Config((pname, site, here) => {
+  case NumNodes => n
+})
+
+class SupernodeFireSimRocketChipConfig extends Config(
+  new WithNumNodes(4) ++
+  new WithExtMemSize(0x200000000L) ++ // 8GB
+  new FireSimRocketChipConfig)
+
+class SupernodeFireSimRocketChipSingleCoreConfig extends Config(
+  new WithNumNodes(4) ++
+  new WithExtMemSize(0x200000000L) ++ // 8GB
+  new FireSimRocketChipSingleCoreConfig)
+
+class SupernodeSixNodeFireSimRocketChipSingleCoreConfig extends Config(
+  new WithNumNodes(6) ++
+  new WithExtMemSize(0x40000000L) ++ // 1GB
+  new FireSimRocketChipSingleCoreConfig)
+
+class SupernodeEightNodeFireSimRocketChipSingleCoreConfig extends Config(
+  new WithNumNodes(8) ++
+  new WithExtMemSize(0x40000000L) ++ // 1GB
+  new FireSimRocketChipSingleCoreConfig)
+
+class SupernodeFireSimRocketChipDualCoreConfig extends Config(
+  new WithNumNodes(4) ++
+  new WithExtMemSize(0x200000000L) ++ // 8GB
+  new FireSimRocketChipDualCoreConfig)
+
+class SupernodeFireSimRocketChipQuadCoreConfig extends Config(
+  new WithNumNodes(4) ++
+  new WithExtMemSize(0x200000000L) ++ // 8GB
+  new FireSimRocketChipQuadCoreConfig)
+
+class SupernodeFireSimRocketChipHexaCoreConfig extends Config(
+  new WithNumNodes(4) ++
+  new WithExtMemSize(0x200000000L) ++ // 8GB
+  new FireSimRocketChipHexaCoreConfig)
+
+class SupernodeFireSimRocketChipOctaCoreConfig extends Config(
+  new WithNumNodes(4) ++
+  new WithExtMemSize(0x200000000L) ++ // 8GB
+  new FireSimRocketChipOctaCoreConfig)
+
